requirements.txt
Django==5.1.15
djangorestframework==3.15.2
djangorestframework-simplejwt==5.5.1
celery==5.4.0
redis==5.2.1
psycopg2-binary==2.9.10
django-celery-results==2.5.1
gunicorn==23.0.0
python-decouple==3.8
dj-database-url==2.3.0
django-cors-headers==4.6.0
django-htmx==1.21.0
pytest==8.3.5
pytest-django==4.9.0
pytest-cov==6.1.1
factory-boy==3.3.1
drf-spectacular==0.28.0
whitenoise==6.8.2
